Neuroconductor Installer

Install function for neuroconductor packages

neuro_install(
  repo,
  release = c("stable", "current"),
  release_repo = latest_neuroc_release(),
  upgrade_dependencies = FALSE,
  type = getOption("pkgType"),
  ...
)

neuroc_install(...)

neurocLite(...)

Arguments

repo
Package name in neuroconductor

release
Stable or current (development) versions/branches

release_repo
Repository for release repository, passed to install.packages. If release_repo = "github", then it will install using GitHub. If you set this using make_release_version or specify the URL directly, it will override release option.

upgrade_dependencies
Should dependencies be updated? passed to install if using release_repo = "github"

type
character, indicating the type of package to download and install, passed to install.packages.

...
additional arguments passed to install.packages or install_github if release_repo = "github"

Value

Result from install.packages or install_github

Examples

# \donttest{
   tlib = tempfile()
   dir.create(tlib, showWarnings = FALSE)
   system.time({
   install.packages("oro.asl",
   lib = tlib,
   repos = "https://neuroconductor.org/releases/2019/12/",
   verbose = TRUE)
   })
#> system (cmd0): /opt/R/4.0.5/lib/R/bin/R CMD INSTALL
#> foundpkgs: oro.asl, /tmp/RtmpTWNh48/downloaded_packages/oro.asl_0.1.1.tar.gz
#> files: /tmp/RtmpTWNh48/downloaded_packages/oro.asl_0.1.1.tar.gz
#> 1): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/oro.asl_0.1.1.tar.gz''
#>    user  system elapsed 
#>   1.222   0.189   2.583 
   repos = getOption("repos")
   print(repos)
#>                          CRAN 
#> "https://cloud.r-project.org" 
   #if (repos["CRAN"] == "@CRAN@") {
   #    repos["CRAN"] = "https://cloud.r-project.org"
   #    options(repos = repos)
   #}
   options(repos = NULL)
   print(getOption("repos"))
#> NULL
   neuro_install("oro.asl", lib = tlib,
   release_repo = "https://neuroconductor.org/releases/2019/12")
   options(repos = repos)
 # }
# \donttest{
   options(repos = "http://cran.r-project.org")
   neuro_install("cifti", type = "source", lib = tlib, verbose = TRUE)
#> system (cmd0): /opt/R/4.0.5/lib/R/bin/R CMD INSTALL
#> also installing the dependencies ‘bitops’, ‘abind’, ‘RNifti’, ‘R.oo’, ‘R.methodsS3’, ‘oro.nifti’, ‘gifti’, ‘R.utils’
#> foundpkgs: bitops, abind, RNifti, R.oo, R.methodsS3, oro.nifti, gifti, R.utils, cifti, /tmp/RtmpTWNh48/downloaded_packages/bitops_1.0-7.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/abind_1.4-5.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/RNifti_1.3.1.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/R.oo_1.24.0.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/R.methodsS3_1.8.1.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/oro.nifti_0.11.0.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/gifti_0.8.0.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/R.utils_2.10.1.tar.gz, /tmp/RtmpTWNh48/downloaded_packages/cifti_0.4.6.0.tar.gz
#> files: /tmp/RtmpTWNh48/downloaded_packages/bitops_1.0-7.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/abind_1.4-5.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/RNifti_1.3.1.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/R.oo_1.24.0.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/R.methodsS3_1.8.1.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/oro.nifti_0.11.0.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/gifti_0.8.0.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/R.utils_2.10.1.tar.gz, 
#> 	/tmp/RtmpTWNh48/downloaded_packages/cifti_0.4.6.0.tar.gz
#> 1): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/bitops_1.0-7.tar.gz''
#> 2): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/abind_1.4-5.tar.gz''
#> 3): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/RNifti_1.3.1.tar.gz''
#> 4): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/R.methodsS3_1.8.1.tar.gz''
#> 5): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/R.oo_1.24.0.tar.gz''
#> 6): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/oro.nifti_0.11.0.tar.gz''
#> 7): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/R.utils_2.10.1.tar.gz''
#> 8): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/gifti_0.8.0.tar.gz''
#> 9): succeeded '/opt/R/4.0.5/lib/R/bin/R CMD INSTALL -l '/tmp/RtmpTWNh48/file8bff3dc05a11' '/tmp/RtmpTWNh48/downloaded_packages/cifti_0.4.6.0.tar.gz''
   neuro_install("cifti",
   release_repo = latest_neuroc_release(),
   lib = tlib)

   neuro_install("cifti", release_repo = "github")
#> Using bundled GitHub PAT. Please add your own PAT to the env var `GITHUB_PAT`
#> Downloading GitHub repo neuroconductor/cifti@a7e91f2f3d3e678bd178f132fe0614499802213d
#> Downloading GitHub repo neuroconductor/gifti@e3ff5e0beb859ef5b27242365b9d4487873adb9e
#> Installing 3 packages: R.methodsS3, R.oo, R.utils
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#>      checking for file ‘/tmp/RtmpTWNh48/remotes8bff275953b/neuroconductor-gifti-e3ff5e0/DESCRIPTION’ ...  ✔  checking for file ‘/tmp/RtmpTWNh48/remotes8bff275953b/neuroconductor-gifti-e3ff5e0/DESCRIPTION’
#>   ─  preparing ‘gifti’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘gifti_0.7.6.tar.gz’
#>      
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Downloading GitHub repo neuroconductor/oro.nifti@bd06b05d2a090c743e51d745a826e8b77af278f0
#> Downloading GitHub repo neuroconductor/RNifti@929106d1b17d6c09c8b001b5e39d448ed9ef635f
#>      checking for file ‘/tmp/RtmpTWNh48/remotes8bff3482a3b/neuroconductor-RNifti-929106d/DESCRIPTION’ ...  ✔  checking for file ‘/tmp/RtmpTWNh48/remotes8bff3482a3b/neuroconductor-RNifti-929106d/DESCRIPTION’
#>   ─  preparing ‘RNifti’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#> ─  cleaning src
#>   ─  running ‘cleanup’
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>   ─  building ‘RNifti_1.3.1.tar.gz’
#>      
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Installing 3 packages: rticles, abind, bitops
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Skipping install of 'RNifti' from a github remote, the SHA1 (929106d1) has not changed since last install.
#>   Use `force = TRUE` to force installation
#>      checking for file ‘/tmp/RtmpTWNh48/remotes8bff6ab4cec2/neuroconductor-oro.nifti-bd06b05/DESCRIPTION’ ...  ✔  checking for file ‘/tmp/RtmpTWNh48/remotes8bff6ab4cec2/neuroconductor-oro.nifti-bd06b05/DESCRIPTION’
#>   ─  preparing ‘oro.nifti’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>      Removed empty directory ‘oro.nifti/vignettes’
#>   ─  building ‘oro.nifti_0.10.3.tar.gz’
#>      
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Installing 6 packages: R.methodsS3, R.oo, R.utils, RNifti, abind, bitops
#> Installing packages into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
#> Skipping install of 'gifti' from a github remote, the SHA1 (e3ff5e0b) has not changed since last install.
#>   Use `force = TRUE` to force installation
#> Skipping install of 'oro.nifti' from a github remote, the SHA1 (bd06b05d) has not changed since last install.
#>   Use `force = TRUE` to force installation
#>      checking for file ‘/tmp/RtmpTWNh48/remotes8bff35d0d5a5/neuroconductor-cifti-a7e91f2/DESCRIPTION’ ...  ✔  checking for file ‘/tmp/RtmpTWNh48/remotes8bff35d0d5a5/neuroconductor-cifti-a7e91f2/DESCRIPTION’
#>   ─  preparing ‘cifti’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>    Removed empty directory ‘cifti/docs’
#>   ─  building ‘cifti_0.4.7.tar.gz’
#>      
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)
# }