Neuroconductor Installer

Install function for neuroconductor packages

neuro_install(
  repo,
  release = c("stable", "current"),
  release_repo = latest_neuroc_release(),
  upgrade_dependencies = FALSE,
  type = getOption("pkgType"),
  ...
)

neuroc_install(...)

neurocLite(...)

Arguments

repo
Package name in neuroconductor

release
Stable or current (development) versions/branches

release_repo
Repository for release repository, passed to install.packages. If release_repo = "github", then it will install using GitHub. If you set this using make_release_version or specify the URL directly, it will override release option.

upgrade_dependencies
Should dependencies be updated? passed to install if using release_repo = "github"

type
character, indicating the type of package to download and install, passed to install.packages.

...
additional arguments passed to install.packages or install_github if release_repo = "github"

Value

Result from install.packages or install_github

Examples

# \donttest{
   tlib = tempfile()
   dir.create(tlib, showWarnings = FALSE)
   system.time({
   install.packages("oro.asl",
   lib = tlib,
   repos = "https://neuroconductor.org/releases/2019/12/",
   verbose = TRUE)
   })

#> 
#> The downloaded binary packages are in
#> 	/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages

#>    user  system elapsed 
#>   0.066   0.023   1.264 

   repos = getOption("repos")
   print(repos)

#>                          CRAN 
#> "https://cloud.r-project.org" 

   #if (repos["CRAN"] == "@CRAN@") {
   #    repos["CRAN"] = "https://cloud.r-project.org"
   #    options(repos = repos)
   #}
   options(repos = NULL)
   print(getOption("repos"))

#> NULL

   neuro_install("oro.asl", lib = tlib,
   release_repo = "https://neuroconductor.org/releases/2019/12")

#> 
#> The downloaded binary packages are in
#> 	/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages

   options(repos = repos)
 # }
# \donttest{
   options(repos = "http://cran.r-project.org")
   neuro_install("cifti", type = "source", lib = tlib, verbose = TRUE)

#> system (cmd0): /Library/Frameworks/R.framework/Resources/bin/R CMD INSTALL

#> foundpkgs: cifti, /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages/cifti_0.4.6.0.tar.gz

#> files: /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages/cifti_0.4.6.0.tar.gz

#> 1): succeeded '/Library/Frameworks/R.framework/Resources/bin/R CMD INSTALL -l '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpB84aun/filedf604786258' '/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages/cifti_0.4.6.0.tar.gz''

   neuro_install("cifti",
   release_repo = latest_neuroc_release(),
   lib = tlib)

#> 
#> The downloaded binary packages are in
#> 	/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpB84aun/downloaded_packages

   neuro_install("cifti", release_repo = "github")

#> Using github PAT from envvar GITHUB_PAT

#> Downloading GitHub repo neuroconductor/cifti@dd98b8c27731fb1e62235827cd6771d2923adcab

#>      checking for file ‘/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpB84aun/remotesdf6080da08d/neuroconductor-cifti-dd98b8c/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpB84aun/remotesdf6080da08d/neuroconductor-cifti-dd98b8c/DESCRIPTION’
#>   ─  preparing ‘cifti’:
#>      checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>    Removed empty directory ‘cifti/docs’
#>   ─  building ‘cifti_0.4.7.tar.gz’
#>      
#> 

#> Installing package into ‘/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpqA9394/temp_libpathdf2f643510d6’
#> (as ‘lib’ is unspecified)

# }